+++ scripts/tests/e2e-wi0316-employee-payslips-locale-dynamic-ui-gap-fix-phase5.test.ts
@@ -8,6 +8,13 @@ function readUtf8(...parts: string[]) {
 
 async function run() {
   const payslipPage = readUtf8("src", "app", "employee", "payslips", "page.tsx");
+  const payslipLocaleHelpers = readUtf8(
+    "src",
+    "app",
+    "employee",
+    "payslips",
+    "page-locale-helpers.ts"
+  );
   const workItem = readUtf8(
     "work-items",
     "WI-0316-employee-payslips-locale-dynamic-ui-gap-fix-phase5.md"
@@ -18,14 +25,20 @@ async function run() {
   assert.match(payslipPage, /const \{ locale \} = useI18n\(\);/);
   assert.match(payslipPage, /const isKoLocale = locale === "ko";/);
   assert.match(payslipPage, /const runtimeLocale = isKoLocale \? "ko-KR" : "en-US";/);
-  assert.match(payslipPage, /const searchSortCopy = useMemo\(/);
-  assert.match(payslipPage, /title: "명세서 검색\/정렬"/);
-  assert.match(payslipPage, /title: "Payslip Search\/Sort"/);
+  assert.match(payslipPage, /from "@\/app\/employee\/payslips\/page-locale-helpers"/);
+  assert.match(payslipPage, /const searchSortCopy = useMemo\(\(\) => resolvePayslipSearchSortCopy\(isKoLocale\), \[isKoLocale\]\);/);
   assert.match(payslipPage, /<h2>\{searchSortCopy\.title\}<\/h2>/);
   assert.match(payslipPage, /<p className="small">\{searchSortCopy\.description\}<\/p>/);
   assert.match(payslipPage, /new Date\(\)\.toLocaleString\(runtimeLocale\)/);
-  assert.match(payslipPage, /parsed\.toLocaleString\(resolveRuntimeLocale\(\)\)/);
-  assert.match(payslipPage, /parsed\.toLocaleDateString\(resolveRuntimeLocale\(\)\)/);
+
+  assert.match(payslipLocaleHelpers, /export function resolveRuntimeLocale\(\)/);
+  assert.match(payslipLocaleHelpers, /export function formatDateTime\(value: string \| null\)/);
+  assert.match(payslipLocaleHelpers, /export function formatDateOnly\(value: string \| null\)/);
+  assert.match(payslipLocaleHelpers, /parsed\.toLocaleString\(resolveRuntimeLocale\(\)\)/);
+  assert.match(payslipLocaleHelpers, /parsed\.toLocaleDateString\(resolveRuntimeLocale\(\)\)/);
+  assert.match(payslipLocaleHelpers, /title: "명세서 검색\/정렬"/);
+  assert.match(payslipLocaleHelpers, /title: "Payslip Search\/Sort"/);
+
   assert.doesNotMatch(payslipPage, /toLocaleString\("ko-KR"\)/);
   assert.doesNotMatch(payslipPage, /toLocaleDateString\("ko-KR"\)/);
